Split landmark featurization into two steps
